removed memcpy every frame but using the pointer

+++ colorgradient.hpp
@@ -1,10 +1,11 @@
 #include <vector>
-#include <glm/vec3.hpp>
+#include <cmath>
+#include <glm/glm.hpp>
 
 class ColorGradient {
     public:
         std::vector<glm::vec3> colors;    
-        glm::vec3 sample(double time);
+        glm::vec3 sample(float time);
 
     private:
 };
